requirements.txt
anyio==4.0.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
Django==4.2.5
django-ckeditor==6.7.0
django-hide==0.0.2
django-js-asset==2.1.0
django-otp==1.2.3
exceptiongroup==1.2.2
filehash==0.2.dev1
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.146.0
google-auth==2.23.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.60.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
idna==3.4
oauthlib==3.2.2
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pypng==0.20220715.0
python-decouple==3.8
python-telegram-bot==20.6
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
